move snapshot operations to snapshot service
