+++ backend/src/taxonomy_builder/services/publishing_service.py
@@ -1,9 +1,8 @@
-"""Service for the publishing workflow: validation, diff, and orchestration."""
+"""Service for the publishing workflow orchestration."""
 
 from datetime import datetime
 from uuid import UUID
 
-from pydantic import ValidationError as PydanticValidationError
 from sqlalchemy import select
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -16,16 +15,16 @@ from taxonomy_builder.schemas.publishing import (
     UpdateDraftRequest,
 )
 from taxonomy_builder.schemas.snapshot import (
-    DiffItem,
     DiffResult,
-    FieldChange,
-    ModifiedItem,
     SnapshotVocabulary,
-    ValidationError,
     ValidationResult,
 )
 from taxonomy_builder.services.project_service import ProjectService
-from taxonomy_builder.services.snapshot_service import SnapshotService
+from taxonomy_builder.services.snapshot_service import (
+    SnapshotService,
+    compute_diff,
+    validate_snapshot,
+)
 
 
 class ValidationFailedError(Exception):
@@ -80,190 +79,6 @@ class PublishingService:
         self._project_service = project_service
         self._snapshot_service = snapshot_service
 
-    @staticmethod
-    def validate_snapshot(snapshot: SnapshotVocabulary) -> ValidationResult:
-        """Validate a snapshot is ready to publish.
-
-        Runs Pydantic validators on the snapshot data, collecting all errors.
-        """
-        try:
-            SnapshotVocabulary.model_validate(snapshot.model_dump(mode="json"))
-            return ValidationResult(valid=True, errors=[])
-        except PydanticValidationError as e:
-            errors = []
-            for err in e.errors():
-                ctx = err.get("ctx", {})
-                entity_id_str = ctx.get("entity_id")
-                errors.append(
-                    ValidationError(
-                        code=err["type"],
-                        message=err["msg"],
-                        entity_type=ctx.get("entity_type"),
-                        entity_id=UUID(entity_id_str) if entity_id_str else None,
-                        entity_label=ctx.get("entity_label"),
-                    )
-                )
-            return ValidationResult(valid=False, errors=errors)
-
-    @staticmethod
-    def _field_changes(prev, curr, exclude: set[str]) -> list[FieldChange]:
-        prev_data = prev.model_dump(exclude=exclude)
-        curr_data = curr.model_dump(exclude=exclude)
-        return [
-            FieldChange(field=f, old=str(prev_data[f]), new=str(curr_data[f]))
-            for f in prev_data
-            if prev_data[f] != curr_data[f]
-        ]
-
-    @staticmethod
-    def compute_diff(
-        previous: SnapshotVocabulary | None,
-        current: SnapshotVocabulary,
-    ) -> DiffResult:
-        """Diff two snapshots, returning added/modified/removed items."""
-        prev_schemes = {s.id: s for s in previous.concept_schemes} if previous else {}
-        curr_schemes = {s.id: s for s in current.concept_schemes}
-        prev_props = {p.id: p for p in previous.properties} if previous else {}
-        curr_props = {p.id: p for p in current.properties}
-        prev_classes = {c.uri: c for c in previous.classes} if previous else {}
-        curr_classes = {c.uri: c for c in current.classes}
-
-        # Categorise changes
-        added_schemes = [
-            curr_schemes[scheme_id] for scheme_id in curr_schemes.keys() - prev_schemes.keys()
-        ]
-        removed_schemes = [
-            prev_schemes[scheme_id] for scheme_id in prev_schemes.keys() - curr_schemes.keys()
-        ]
-        modified_schemes = [
-            (
-                prev_schemes[scheme_id],
-                curr_schemes[scheme_id],
-                {concept.id: concept for concept in prev_schemes[scheme_id].concepts},
-                {concept.id: concept for concept in curr_schemes[scheme_id].concepts},
-            )
-            for scheme_id in prev_schemes.keys() & curr_schemes.keys()
-        ]
-        added_properties = [curr_props[pid] for pid in curr_props.keys() - prev_props.keys()]
-        removed_properties = [prev_props[pid] for pid in prev_props.keys() - curr_props.keys()]
-        modified_properties = [
-            (prev_props[pid], curr_props[pid]) for pid in prev_props.keys() & curr_props.keys()
-        ]
-        added_classes = [curr_classes[uri] for uri in curr_classes.keys() - prev_classes.keys()]
-        removed_classes = [prev_classes[uri] for uri in prev_classes.keys() - curr_classes.keys()]
-
-        added = (
-            # New schemes
-            [
-                DiffItem(id=scheme.id, label=scheme.title, entity_type="scheme")
-                for scheme in added_schemes
-            ]
-            # Concepts in new schemes
-            + [
-                DiffItem(id=concept.id, label=concept.pref_label, entity_type="concept")
-                for scheme in added_schemes
-                for concept in scheme.concepts
-            ]
-            # New concepts in existing schemes
-            + [
-                DiffItem(
-                    id=concept_id, label=curr_concepts[concept_id].pref_label, entity_type="concept"
-                )
-                for _, _, prev_concepts, curr_concepts in modified_schemes
-                for concept_id in curr_concepts.keys() - prev_concepts.keys()
-            ]
-            # New properties
-            + [
-                DiffItem(id=property.id, label=property.label, entity_type="property")
-                for property in added_properties
-            ]
-            # New classes
-            + [DiffItem(uri=cls.uri, label=cls.label, entity_type="class") for cls in added_classes]
-        )
-
-        removed = (
-            # Removed schemes
-            [
-                DiffItem(id=scheme.id, label=scheme.title, entity_type="scheme")
-                for scheme in removed_schemes
-            ]
-            # Concepts in removed schemes
-            + [
-                DiffItem(id=concept.id, label=concept.pref_label, entity_type="concept")
-                for scheme in removed_schemes
-                for concept in scheme.concepts
-            ]
-            # Removed concepts in existing schemes
-            + [
-                DiffItem(
-                    id=concept_id, label=prev_concepts[concept_id].pref_label, entity_type="concept"
-                )
-                for _, _, prev_concepts, curr_concepts in modified_schemes
-                for concept_id in prev_concepts.keys() - curr_concepts.keys()
-            ]
-            # Removed properties
-            + [
-                DiffItem(id=property.id, label=property.label, entity_type="property")
-                for property in removed_properties
-            ]
-            # Removed classes
-            + [
-                DiffItem(uri=cls.uri, label=cls.label, entity_type="class")
-                for cls in removed_classes
-            ]
-        )
-
-        modified = (
-            # Modified scheme metadata
-            [
-                ModifiedItem(
-                    id=curr_scheme.id,
-                    label=curr_scheme.title,
-                    entity_type="scheme",
-                    changes=changes,
-                )
-                for prev_scheme, curr_scheme, _, _ in modified_schemes
-                if (
-                    changes := PublishingService._field_changes(
-                        prev_scheme, curr_scheme, {"id", "concepts"}
-                    )
-                )
-            ]
-            # Modified concepts in existing schemes
-            + [
-                ModifiedItem(
-                    id=concept_id,
-                    label=curr_concepts[concept_id].pref_label,
-                    entity_type="concept",
-                    changes=changes,
-                )
-                for _, _, prev_concepts, curr_concepts in modified_schemes
-                for concept_id in prev_concepts.keys() & curr_concepts.keys()
-                if (
-                    changes := PublishingService._field_changes(
-                        prev_concepts[concept_id], curr_concepts[concept_id], {"id"}
-                    )
-                )
-            ]
-            # Modified properties
-            + [
-                ModifiedItem(
-                    id=curr_property.id,
-                    label=curr_property.label,
-                    entity_type="property",
-                    changes=changes,
-                )
-                for prev_property, curr_property in modified_properties
-                if (
-                    changes := PublishingService._field_changes(
-                        prev_property, curr_property, {"id"}
-                    )
-                )
-            ]
-        )
-
-        return DiffResult(added=added, modified=modified, removed=removed)
-
     async def publish(
         self,
         project_id: UUID,
@@ -272,7 +87,7 @@ class PublishingService:
     ) -> PublishedVersion:
         """Publish a new version (or create a draft) of a project."""
         snapshot = await self._snapshot_service.build_snapshot(project_id)
-        validation = self.validate_snapshot(snapshot)
+        validation = validate_snapshot(snapshot)
         if not validation.valid:
             raise ValidationFailedError(validation)
 
@@ -319,7 +134,7 @@ class PublishingService:
             raise NotADraftError(version_id)
 
         snapshot = await self._snapshot_service.build_snapshot(project_id)
-        validation = self.validate_snapshot(snapshot)
+        validation = validate_snapshot(snapshot)
         if not validation.valid:
             raise ValidationFailedError(validation)
 
@@ -359,7 +174,7 @@ class PublishingService:
     async def preview(self, project_id: UUID) -> PublishPreview:
         """Build a publish preview: validation, content summary, and diff."""
         snapshot = await self._snapshot_service.build_snapshot(project_id)
-        validation = self.validate_snapshot(snapshot)
+        validation = validate_snapshot(snapshot)
 
         content_summary = ContentSummary(
             schemes=len(snapshot.concept_schemes),
@@ -371,7 +186,7 @@ class PublishingService:
         diff = None
         if latest:
             prev_snapshot = SnapshotVocabulary.model_validate(latest.snapshot)
-            diff = self.compute_diff(prev_snapshot, snapshot)
+            diff = compute_diff(prev_snapshot, snapshot)
 
         suggested = self._suggest_version(latest.version if latest else None, diff)
 
@@ -394,7 +209,7 @@ class PublishingService:
             raise NotADraftError(version_id)
 
         snapshot = await self._snapshot_service.build_snapshot(project_id)
-        validation = self.validate_snapshot(snapshot)
+        validation = validate_snapshot(snapshot)
         if not validation.valid:
             raise ValidationFailedError(validation)
