+++ backend/tests/test_services/test_snapshot_validation.py
@@ -8,7 +8,7 @@ from taxonomy_builder.schemas.snapshot import (
     SnapshotScheme,
     SnapshotVocabulary,
 )
-from taxonomy_builder.services.publishing_service import PublishingService
+from taxonomy_builder.services.snapshot_service import validate_snapshot
 
 
 def _project_meta() -> SnapshotProjectMetadata:
@@ -51,14 +51,14 @@ class TestValidProject:
     def test_valid_project(self) -> None:
         concept = _concept("Term A")
         scheme = _scheme("Test Scheme", concepts=[concept])
-        result = PublishingService.validate_snapshot(_vocab(scheme))
+        result = validate_snapshot(_vocab(scheme))
         assert result.valid is True
         assert result.errors == []
 
 
 class TestNoSchemes:
     def test_no_schemes(self) -> None:
-        result = PublishingService.validate_snapshot(_vocab())
+        result = validate_snapshot(_vocab())
         assert result.valid is False
         assert any(e.code == "no_schemes" for e in result.errors)
 
@@ -66,7 +66,7 @@ class TestNoSchemes:
 class TestNoConceptsInScheme:
     def test_scheme_with_no_concepts(self) -> None:
         scheme = _scheme("Empty Scheme", concepts=[])
-        result = PublishingService.validate_snapshot(_vocab(scheme))
+        result = validate_snapshot(_vocab(scheme))
         assert result.valid is False
         assert any(e.code == "scheme_no_concepts" for e in result.errors)
 
@@ -75,7 +75,7 @@ class TestSchemeMissingUri:
     def test_scheme_missing_uri(self) -> None:
         scheme_id = uuid4()
         scheme = _scheme("No URI", id=scheme_id, uri=None, concepts=[_concept()])
-        result = PublishingService.validate_snapshot(_vocab(scheme))
+        result = validate_snapshot(_vocab(scheme))
         assert result.valid is False
         uri_errors = [e for e in result.errors if e.code == "scheme_missing_uri"]
         assert len(uri_errors) == 1
@@ -86,7 +86,7 @@ class TestSchemeMissingUri:
 class TestConceptMissingPrefLabel:
     def test_whitespace_only_label(self) -> None:
         scheme = _scheme(concepts=[_concept("   ")])
-        result = PublishingService.validate_snapshot(_vocab(scheme))
+        result = validate_snapshot(_vocab(scheme))
         assert result.valid is False
         assert any(e.code == "concept_missing_pref_label" for e in result.errors)
 
@@ -95,7 +95,7 @@ class TestCollectsAllErrors:
     def test_multiple_errors_collected(self) -> None:
         """Validation returns all errors, not just the first one."""
         scheme = _scheme("Bad", uri=None, concepts=[_concept("  ")])
-        result = PublishingService.validate_snapshot(_vocab(scheme))
+        result = validate_snapshot(_vocab(scheme))
         assert result.valid is False
         codes = {e.code for e in result.errors}
         assert "scheme_missing_uri" in codes
@@ -107,7 +107,7 @@ class TestMixedValidity:
         good = _scheme("Good", concepts=[_concept("Valid")])
         bad_id = uuid4()
         bad = _scheme("Bad", id=bad_id, uri=None, concepts=[_concept("Also Valid")])
-        result = PublishingService.validate_snapshot(_vocab(good, bad))
+        result = validate_snapshot(_vocab(good, bad))
         assert result.valid is False
         uri_errors = [e for e in result.errors if e.code == "scheme_missing_uri"]
         assert len(uri_errors) == 1
